Split parsers.py into a package
