+++ habitat/parsers/habitat_base.py
@@ -1,102 +1,22 @@
 from __future__ import print_function
 import anytree
-from .errors import DuplicateJsonError, _IgnoredVersionError
-from .solvers import Solver
-from future.utils import with_metaclass
-import json
 import logging
 import os
-from packaging.requirements import Requirement
-from packaging.specifiers import SpecifierSet
-from packaging.version import Version, InvalidVersion
-from pprint import pformat
-import sys
 import subprocess
+import sys
 import tabulate
 
+from future.utils import with_metaclass
+from packaging.version import Version
+from pprint import pformat
+
+from . import HabitatMeta, NotSet, habitat_property
+from .. import json
+from ..errors import DuplicateJsonError
+from ..solvers import Solver
+
+
 logger = logging.getLogger(__name__)
-
-
-class NotSet(object):
-    """The data for this property is not currently set."""
-
-    def __bool__(self):
-        """NotSet should be treated as False when booled Python 3"""
-        return False
-
-    def __nonzero__(self):
-        """NotSet should be treated as False when booled Python 2"""
-        return False
-
-    def __str__(self):
-        return "NotSet"
-
-
-# Make this a singleton so it works like a boolean False for if statements.
-NotSet = NotSet()
-
-
-class _HabitatProperty(property):
-    """The @property decorator that can be type checked by the `HabitatMeta` metaclass
-
-    Any properties using this decorator will have their name added to `_properties`.
-    Don't use this class directly, use the habitat_property decorator instead.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super(_HabitatProperty, self).__init__(*args, **kwargs)
-        # Copy from the class, and store it on the instance. This makes it so we
-        # can use isinstance checks against this class.
-        cls = type(self)
-        self.group = cls.group
-        self.process_order = cls.process_order
-        self.verbosity = cls.verbosity
-
-    def sort_key(self):
-        """Used to provide consistent and controlled when dynamically processing
-        _HabitatProperties.
-        """
-        return (self.process_order, self.fget.__name__)
-
-
-def habitat_property(verbosity=0, group=1, process_order=100):
-    """Decorate this function as a property and configure how it shows up in the dump.
-
-    Args:
-        verbosity (int, optional): Specify the minimum verbosity required to show.
-        group (int, optional): Controls how dump sorts the results. The sort uses
-            (group, name) as its sort key, so common group values are sorted
-            alphabetically. This should rarely be used and only 0 or 1 should be used
-            to reduce the complexity of finding the property you are looking for.
-        process_order (int, optional): Fine control over the order properties are
-            processed in. The sort uses (process_order, name) as its sort key, so
-            common process_order values are sorted alphabetically. This should only
-            be needed in specific cases. This is used by ``_HabitatProperty.sort_key``.
-    """
-    # Store the requested values to the class so __init__ can copy them into its
-    # instance when initialized by the decoration process.
-    _HabitatProperty.group = group
-    _HabitatProperty.process_order = process_order
-    _HabitatProperty.verbosity = verbosity
-    return _HabitatProperty
-
-
-class HabitatMeta(type):
-    """Scans for HabitatProperties and adds their name to the `_properties` dict."""
-
-    def __new__(cls, name, bases, dct):
-        desc = {}
-        # Include any _properties that are added by base classes we are inheriting.
-        for base in bases:
-            if hasattr(base, "_properties"):
-                desc.update(base._properties)
-
-        # Add any new _properties defined on this class
-        for k, v in dct.items():
-            if isinstance(v, _HabitatProperty):
-                desc[k] = v
-        dct["_properties"] = desc
-        return type.__new__(cls, name, bases, dct)
 
 
 class HabitatBase(with_metaclass(HabitatMeta, anytree.NodeMixin)):
@@ -521,6 +441,8 @@ class HabitatBase(with_metaclass(HabitatMeta, anytree.NodeMixin)):
 
     def reduced(self, resolver=None, uri=None):
         """Returns a new instance with the final settings applied respecting inheritance"""
+        from . import FlatConfig
+
         return FlatConfig(self, resolver, uri=uri)
 
     @classmethod
@@ -710,241 +632,3 @@ class HabitatBase(with_metaclass(HabitatMeta, anytree.NodeMixin)):
         if launch_script:
             with open(launch_script, "w") as fle:
                 fle.write(shell["launch"].format(path=config_script))
-
-
-class Distro(HabitatBase):
-    def latest_version(self, specifier):
-        """Returns the newest version available matching the specifier"""
-        versions = self.matching_versions(specifier)
-        try:
-            version = max(versions)
-        except ValueError:
-            raise Exception('Unable to find a valid version for "{}"'.format(specifier))
-        return self.versions[version]
-
-    def matching_versions(self, specification):
-        """Returns a list of versions available matching the version specification.
-        See `packaging.requirements` for details on valid requirements, but it
-        should be the same as pip requirements.
-        """
-        if isinstance(specification, Requirement):
-            specifier = specification.specifier
-        elif isinstance(specification, SpecifierSet):
-            specifier = specification
-        else:
-            specifier = Requirement(specification).specifier
-        return specifier.filter(
-            self.versions.keys(), prereleases=self.resolver.prereleases
-        )
-
-    @property
-    def versions(self):
-        """A dict of available distro versions"""
-        return {c.version: c for c in self.children}
-
-
-class DistroVersion(HabitatBase):
-    _context_method = "name"
-    _placeholder = Distro
-
-    def _init_variables(self):
-        super(DistroVersion, self)._init_variables()
-        self.aliases = NotSet
-
-    @habitat_property()
-    def aliases(self):
-        return self._aliases
-
-    @aliases.setter
-    def aliases(self, aliases):
-        self._aliases = aliases
-
-    def load(self, filename):
-        # Fill in the DistroVersion specific settings before calling super
-        data = self._load(filename)
-        self.aliases = data.get("aliases", NotSet)
-
-        # The version can be stored in several ways to make deployment and dev easier
-        if "version" in data:
-            self.version = data["version"]
-        elif os.path.exists(os.path.join(self.dirname, ".habitat_version.txt")):
-            self.version = (
-                open(os.path.join(self.dirname, ".habitat_version.txt")).read().strip()
-            )
-        else:
-            # If version is not defined in json data extract it from the parent
-            # directory name. This allows for simpler distribution without needing
-            # to modify version controlled files.
-            try:
-                self.version = os.path.basename(self.dirname)
-            except InvalidVersion:
-                """The parent directory was not a valid version, attempt to get a
-                version using setuptools_scm.
-                """
-                from setuptools_scm import get_version
-
-                try:
-                    self.version = get_version(
-                        root=self.dirname, version_scheme="release-branch-semver"
-                    )
-                except LookupError:
-                    if os.path.basename(self.dirname) in self.resolver.ignored:
-                        # This object is not added to the forest until super is called
-                        raise _IgnoredVersionError(
-                            'Skipping "{}" its dirname is in the ignored list.'.format(
-                                filename
-                            )
-                        )
-                    raise LookupError(
-                        'Habitat was unable to determine the version for "{filename}".\n'
-                        "The version is defined in one of several ways checked in this order:\n"
-                        "1. The version property in `.habitat.json`.\n"
-                        "2. A `.habitat_version.txt` file next to `.habitat.json`.\n"
-                        "3. `.habitat.json`'s parent directory name.\n"
-                        "4. setuptools_scm can get a version from version control.\n"
-                        "The preferred method is #3 for deployed releases. #4 is the "
-                        "preferred method for developers working copies.".format(
-                            filename=self.filename
-                        )
-                    )
-
-        # The name should be the version == specifier.
-        self.distro_name = data.get("name")
-        self.name = u"{}=={}".format(self.distro_name, self.version)
-
-        data = super(DistroVersion, self).load(filename, data=data)
-
-        return data
-
-    @habitat_property()
-    def version(self):
-        return super(DistroVersion, self).version
-
-    @version.setter
-    def version(self, version):
-        # NOTE: super doesn't work for a @property.setter
-        if version and not isinstance(version, Version):
-            version = Version(version)
-        self._version = version
-
-
-class Config(HabitatBase):
-    def _init_variables(self):
-        super(Config, self)._init_variables()
-        self.inherits = NotSet
-
-    @habitat_property(verbosity=2)
-    def inherits(self):
-        return self._inherits
-
-    @inherits.setter
-    def inherits(self, inherits):
-        self._inherits = inherits
-
-    def load(self, filename):
-        data = super(Config, self).load(filename)
-        self.inherits = data.get("inherits", NotSet)
-        return data
-
-    @habitat_property(verbosity=1, group=0)
-    def uri(self):
-        # Mark uri as a HabitatProperty so it is included in _properties
-        return super(Config, self).uri
-
-
-class FlatConfig(Config):
-    def __init__(self, original_node, resolver, uri=NotSet):
-        super(FlatConfig, self).__init__(original_node.forest, resolver)
-        self.original_node = original_node
-        self.filename = original_node.filename
-        self._context = original_node.context
-        self._uri = uri
-        self._versions = None
-        # Copy the properties from the inheritance system
-        self._collect_values(self.original_node)
-
-    def _collect_values(self, node, default=False):
-        logger.debug("Loading node: {} inherits: {}".format(node.name, node.inherits))
-        self._missing_values = False
-        # Use sort_key to ensure the properties are processed in the correct order
-        for attrname in sorted(
-            self._properties, key=lambda i: self._properties[i].sort_key()
-        ):
-            if attrname == "uri":
-                # TODO: Add detection of setters to HabitatProperty and don't set values without setters
-                # There is no setter for uri, setting it now will cause errors in testing
-                continue
-            if getattr(self, attrname) != NotSet:
-                continue
-            value = getattr(node, attrname)
-            if value is NotSet:
-                self._missing_values = True
-            else:
-                setattr(self, attrname, value)
-        if node.inherits and self._missing_values:
-            parent = node.parent
-            if parent:
-                return self._collect_values(parent, default=default)
-            elif not default and "default" in self.forest:
-                # Start processing the default setup
-                default = True
-                default_node = self.resolver.closest_config(node.fullpath)
-                self._collect_values(default_node, default=default)
-
-        return self._missing_values
-
-    @habitat_property()
-    def aliases(self):
-        ret = {}
-        for version in self.versions:
-            if version.aliases:
-                aliases_def = version.aliases.get(self._platform, [])
-                aliases = [a[1] for a in aliases_def]
-
-                for i, alias in enumerate(aliases_def):
-                    ret[alias[0]] = version.format_environment_value(aliases[i])
-
-        return ret
-
-    @property
-    def environment(self):
-        """A resolved set of environment variables that should be applied to
-        configure an environment. Any values containing a empty string indicate
-        that the variable should be unset.
-        """
-        empty = self._environment is None
-        super(FlatConfig, self).environment
-        # Add any environment variables defined by the linked versions
-        if empty:
-            for version in self.versions:
-                self.update_environment(version.environment_config, obj=version)
-
-        return self._environment
-
-    @property
-    def fullpath(self):
-        return self.separator.join([name for name in self.context] + [self.name])
-
-    @habitat_property(verbosity=1)
-    def versions(self):
-        if self._distros is NotSet:
-            return []
-
-        if self._versions is None:
-            self._versions = []
-            reqs = self.resolver.resolve_requirements(self.distros)
-            for req in reqs.values():
-                self._versions.append(self.resolver.find_distro(req))
-
-        return self._versions
-
-
-class Placeholder(HabitatBase):
-    """Provides an parent node for a child if one hasn't been created yet.
-    This node will be replaced in the tree if a node is loaded for this position.
-    """
-
-
-# This is the first place where both HabitatBase and its subclass Placeholder
-# are defined, so this is where we have to set _placeholder.
-HabitatBase._placeholder = Placeholder
