+++ habitat/parsers/meta.py
@@ -1,0 +1,76 @@
+class NotSet(object):
+    """The data for this property is not currently set."""
+
+    def __bool__(self):
+        """NotSet should be treated as False when booled Python 3"""
+        return False
+
+    def __str__(self):
+        return "NotSet"
+
+
+# Make this a singleton so it works like a boolean False for if statements.
+NotSet = NotSet()
+
+
+class _HabitatProperty(property):
+    """The @property decorator that can be type checked by the `HabitatMeta` metaclass
+
+    Any properties using this decorator will have their name added to `_properties`.
+    Don't use this class directly, use the habitat_property decorator instead.
+    """
+
+    def __init__(self, *args, **kwargs):
+        super(_HabitatProperty, self).__init__(*args, **kwargs)
+        # Copy from the class, and store it on the instance. This makes it so we
+        # can use isinstance checks against this class.
+        cls = type(self)
+        self.group = cls.group
+        self.process_order = cls.process_order
+        self.verbosity = cls.verbosity
+
+    def sort_key(self):
+        """Used to provide consistent and controlled when dynamically processing
+        _HabitatProperties.
+        """
+        return (self.process_order, self.fget.__name__)
+
+
+def habitat_property(verbosity=0, group=1, process_order=100):
+    """Decorate this function as a property and configure how it shows up in the dump.
+
+    Args:
+        verbosity (int, optional): Specify the minimum verbosity required to show.
+        group (int, optional): Controls how dump sorts the results. The sort uses
+            (group, name) as its sort key, so common group values are sorted
+            alphabetically. This should rarely be used and only 0 or 1 should be used
+            to reduce the complexity of finding the property you are looking for.
+        process_order (int, optional): Fine control over the order properties are
+            processed in. The sort uses (process_order, name) as its sort key, so
+            common process_order values are sorted alphabetically. This should only
+            be needed in specific cases. This is used by ``_HabitatProperty.sort_key``.
+    """
+    # Store the requested values to the class so __init__ can copy them into its
+    # instance when initialized by the decoration process.
+    _HabitatProperty.group = group
+    _HabitatProperty.process_order = process_order
+    _HabitatProperty.verbosity = verbosity
+    return _HabitatProperty
+
+
+class HabitatMeta(type):
+    """Scans for HabitatProperties and adds their name to the `_properties` dict."""
+
+    def __new__(cls, name, bases, dct):
+        desc = {}
+        # Include any _properties that are added by base classes we are inheriting.
+        for base in bases:
+            if hasattr(base, "_properties"):
+                desc.update(base._properties)
+
+        # Add any new _properties defined on this class
+        for k, v in dct.items():
+            if isinstance(v, _HabitatProperty):
+                desc[k] = v
+        dct["_properties"] = desc
+        return type.__new__(cls, name, bases, dct)
