+++ tests/test_resolver.py
@@ -108,22 +108,22 @@ def test_dump_forest(resolver):
     check = "\n".join(
         (
             "default",
-            "    habitat.parsers.Config('default')",
-            "    |-- habitat.parsers.Config('default/Sc1')",
-            "    +-- habitat.parsers.Config('default/Sc11')",
+            "    habitat.parsers.config.Config('default')",
+            "    |-- habitat.parsers.config.Config('default/Sc1')",
+            "    +-- habitat.parsers.config.Config('default/Sc11')",
             "not_set",
-            "    habitat.parsers.Config('not_set')",
-            "    |-- habitat.parsers.Config('not_set/child')",
-            "    |-- habitat.parsers.Config('not_set/env1')",
-            "    |-- habitat.parsers.Config('not_set/env_path_set')",
-            "    |-- habitat.parsers.Config('not_set/env_path_unset')",
-            "    |-- habitat.parsers.Config('not_set/distros')",
-            "    +-- habitat.parsers.Config('not_set/no_env')",
+            "    habitat.parsers.config.Config('not_set')",
+            "    |-- habitat.parsers.config.Config('not_set/child')",
+            "    |-- habitat.parsers.config.Config('not_set/env1')",
+            "    |-- habitat.parsers.config.Config('not_set/env_path_set')",
+            "    |-- habitat.parsers.config.Config('not_set/env_path_unset')",
+            "    |-- habitat.parsers.config.Config('not_set/distros')",
+            "    +-- habitat.parsers.config.Config('not_set/no_env')",
             "project_a",
-            "    habitat.parsers.Config('project_a')",
-            "    +-- habitat.parsers.Config('project_a/Sc001')",
-            "        |-- habitat.parsers.Config('project_a/Sc001/Animation')",
-            "        +-- habitat.parsers.Config('project_a/Sc001/Rigging')",
+            "    habitat.parsers.config.Config('project_a')",
+            "    +-- habitat.parsers.config.Config('project_a/Sc001')",
+            "        |-- habitat.parsers.config.Config('project_a/Sc001/Animation')",
+            "        +-- habitat.parsers.config.Config('project_a/Sc001/Rigging')",
         )
     )
     assert result == check
